✨ feat: Add support for `webOrigins` in SocketController
This commit adds support for the `webOrigins` property to the `PlugConfig` interface and uses it in `SocketController` to populate the redirect URIs when present.

The change also updates the `CreateClientOptions` interface in SocketService to include a `webOrigins` field, which is used to set the `webOrigins` property on the client if provided.

Lastly, the commit includes some minor formatting changes to the affected files.